Added support for out-of-place install of fluidity

+++ build-ese-fetch.sh
@@ -22,46 +22,50 @@ module load ese-software
 module load ese-fluidity-dev # ese-fluidity-dev/20200227-python3-cmake317
 ################################################################################
 
-FLUIDITY_GIT_URL=https://github.com/fluidityproject/fluidity.git
-FLUIDITY_VERSION="VTK9-fixes"
+# Change from yes to not build fluidity every time we build fetch
+export build_fluidity = "yes"
+if test "x$build_fluidity" = "xyes"; then
+  FLUIDITY_GIT_URL=https://github.com/fluidityproject/fluidity.git
+  FLUIDITY_VERSION="VTK9-fixes"
 
-# Clone fluidity from GitHub into fetch if not already present
-if [ ! -d ${FLUIDITY_DIR} ]; then
-  git clone ${FLUIDITY_GIT_URL} ${FLUIDITY_DIR}
+  # Clone fluidity from GitHub into fetch if not already present
+  if [ ! -d ${FLUIDITY_DIR} ]; then
+    git clone ${FLUIDITY_GIT_URL} ${FLUIDITY_DIR}
+  fi
+
+  # Checkout the branch with the VTK9 fixes
+  cd ${FLUIDITY_DIR}
+  git pull
+  git checkout ${FLUIDITY_VERSION}
+  git reset --hard ${FLUIDITY_VERSION}
+
+  #### CHANGE THE PATCH LOCATION TO YOUR PATCH DIRECTORY
+  #
+  # Local changes against Fluidity master at VTK9-fixes
+  patch -p0 <${PATCH_DIR}/use-local-netcdf.patch
+  patch -p0 <${PATCH_DIR}/vtk-add-libs.patch
+  patch -p0 <${PATCH_DIR}/vtk-add-tools-libs.patch
+  patch -p0 <${PATCH_DIR}/vtk-add-fldecomp-libs.patch
+
+  # Regenerate configure after local changes
+  autoconf
+  pushd libadaptivity
+  autoconf
+  popd
+
+  # A bit of prep
+  libtoolize
+
+  # Configure
+  export LIBS="${LIBS} -L${PETSC_DIR}/lib -lblas -llapack"
+
+  ${FLUIDITY_DIR}/configure --enable-2d-adaptivity
+
+  # Build
+  make all -j 8
 fi
 
-# Checkout the branch with the VTK9 fixes
-pushd ${FETCH_DIR}
-pushd ${FLUIDITY_DIR}
-git pull
-git checkout ${FLUIDITY_VERSION}
-git reset --hard ${FLUIDITY_VERSION}
-
-#### CHANGE THE PATCH LOCATION TO YOUR PATCH DIRECTORY
-#
-# Local changes against Fluidity master at VTK9-fixes
-patch -p0 <${PATCH_DIR}/use-local-netcdf.patch
-patch -p0 <${PATCH_DIR}/vtk-add-libs.patch
-patch -p0 <${PATCH_DIR}/vtk-add-tools-libs.patch
-patch -p0 <${PATCH_DIR}/vtk-add-fldecomp-libs.patch
-
-# Regenerate configure after local changes
-autoconf
-pushd libadaptivity
-autoconf
-popd
-
-# A bit of prep
-libtoolize
-
-# Configure
-export LIBS="${LIBS} -L${PETSC_DIR}/lib -lblas -llapack"
-
-${FLUIDITY_DIR}/configure --enable-2d-adaptivity
-
-# Build
-make all -j 8
-popd # takes us to FETCH_DIR
+cd ${FETCH_DIR}
 
 #### CHANGE THE PATCH LOCATION TO YOUR PATCH DIRECTORY
 #
@@ -73,9 +77,7 @@ patch -p0 <${PATCH_DIR}/vtk-add-libs-fetch.patch
 autoconf
 
 # Configure
-${FETCH_DIR}/configure --enable-2d-adaptivity
+${FETCH_DIR}/configure --enable-2d-adaptivity --with-fluidity-root=${FLUIDITY_DIR}
 
 # Build
 make bin/fetch -j 8
-
-popd
